requirements.txt
bcrypt==4.0.1
extra-streamlit-components==0.1.56
langchain==0.0.297
openai==0.27.8
Pillow==9.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
qdrant_client==1.5.4
streamlit==1.25.0
tiktoken==0.3.3
unstructured==0.7.12